+++ server/src/tests/cases/compiler/namedObject1.c
@@ -1,0 +1,12 @@
+test() {
+
+    object "object.c" ob;
+    ob->query_number();
+
+}
+
+// make sure LDMud named objects are parsed
+// and imports resolved
+
+// @driver: ldmud
+// @files: object.c