+++ server/src/tests/cases/compiler/namedObject3.c
@@ -1,0 +1,10 @@
+test() {
+
+    object "notfound.c" ob;    
+
+}
+
+// named objects should report a parser error in FluffOS
+
+// @driver: fluffos
+// @errors: 3